Remove second CC button in admin ticket reply form
The KSD_Admin contructor seems to be called twice as a result
add_tinymce_cc_button is also called twice. This results in the
CC button being displayed twice. The current work around is to
first check if the ksd_cc_button has already been added to the
buttons array but adding it. The issue of multiple calls to the
KSD_Admin construct persists and should be looked into.